requirements.txt
PyGithub==2.3.0
streamlit==1.35.0
plotly==5.22.0
pandas==2.2.2
scikit-learn==1.5.0
textblob==0.18.0
wordcloud==1.9.3
matplotlib==3.9.0
python-dotenv==1.0.1
nltk==3.8.1
